+++ src/components/Checkmark.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { Component, Fragment } from "react";
 
 import classnames from "../classnames";
 
@@ -10,7 +10,7 @@ class Checkmark extends Component {
   };
 
   render() {
-    const { className, checked, onClick } = this.props;
+    const { children, className, checked, onClick } = this.props;
 
     return (
       <button
@@ -25,6 +25,7 @@ class Checkmark extends Component {
           <circle cx="26" cy="26" r="25" fill="none" />
           <path fill="none" d="M14.1 27.2l7.1 7.2 16.7-16.8" />
         </svg>
+        {children && <Fragment>{" "}{children}</Fragment>}
       </button>
     );
   }